Added AssetPack CRD and Logout function

+++ src/services/AssetService.ts
@@ -15,6 +15,16 @@ export default class AssetService implements Service {
     return await AssetPack.find({});
   }
 
+  public async getMyAssets(user: User): Promise<AssetPack[]> {
+    try {
+      console.log(user);
+      const assets = await AssetPack.find({ createdBy: user }).lean();
+      return assets;
+    } catch (error) {
+      throw new ServerException(Errors.SERVER_ERROR);
+    }
+  }
+
   public async getAssetPack(id: string): Promise<AssetPack> {
     let assetPack: AssetPack | null;
     try {
@@ -30,21 +40,22 @@ export default class AssetService implements Service {
     return assetPack;
   }
 
-  public async createAssetPack(options: { title: string; description: string; creator: User }): Promise<AssetPack> {
-    const { title, description, creator } = options;
+  public async createAssetPack(options: { title: string; creator: User }): Promise<AssetPack> {
+    const { title, creator } = options;
     try {
-      const assetPack = new AssetPack({ title, description, createdBy: creator, assets: [] });
+      const assetPack = new AssetPack({ title, createdBy: creator._id });
       await assetPack.save();
 
       return assetPack;
     } catch (error) {
+      console.log(error);
       throw new ServerException(Errors.SERVER_ERROR);
     }
   }
 
-  public async deleteAssetPack(id: string): Promise<void> {
+  public async deleteAssetPack(id: string, creator: User): Promise<void> {
     try {
-      await AssetPack.deleteOne({ _id: new ObjectId(id) });
+      await AssetPack.deleteOne({ _id: new ObjectId(id), createdBy: creator });
     } catch (error) {
       throw new ServerException(Errors.SERVER_ERROR);
     }
